+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.15] - 2021-04-05
+### Added
+- Many log message to help debugging if GrDF site changes something : pygazpar.log
+- Take a screenshot of the corresponding page where a selenium command fails : error_screenshot.png.
+
 ## [0.1.14] - 2020-09-23
 ### Fixed
 - GrDF survey popup has to be closed at the home page.